This is a response to two SCSB observations about CRDS:
1. Interactive users do in fact change data models being fed to CRDS, so
caching data models information based on filename is dangerous and can have
erroneous outcomes.   Caching works great... except for those users who
interactively change CRDS datamodel matching parameters without creating
a new filename.

2. The prefetch really serves no purpose when operating within an individual
Step since individual calls to get_reference_file() typically immediately
follow the Step prefetch.

There are two aspects to this update:

1. Refactor Pipeline and Step to move prefetch to Pipeline only

Moved reference file precaching from Step to Pipeline since precaching in Step
is of limited use and adds an extra CRDS operation and potential datamodel read
for each Step.  This leaves the pipeline level prefetch and individual calls to
get_reference_file() within each Step.  Note that if the prefetch results were
cached then get_reference_file() could effectively be avoided...  but caching
invariably leaves behind land mines for interactive users.   The pipeline level
prefetch is essentially one operation for all Steps and continues to serve the
purpose of broadly checking for errors and ensuring file availability at the
start of a Pipeline.

2. Disable the (dangerous) caching in the crds_client module.

Commented out active areas of CRDS client header + results caching effectively
disabling caching in crds_client w/o fully removing the code or refactoring
much.
